+++ components/layout/Footer/index.js
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 import { linksNav, linksSocial } from '@/utils/links';
-import NewsletterSubscribe from '@/components/mailchimp/NewsletterSubscribe';
+import NewsletterSubscribe from '@/components/NewsletterSubscribe';
 import Container from '@/components/containers/Container';
 import S from './styles';
 
@@ -50,11 +50,20 @@ export default function Footer() {
           </div>
         </S.NavSocialsContainer>
       </S.Inner>
-      <Container>
+      <S.BottomText>
         <S.Copyright>
           © Web Dev Path {new Date().getFullYear()}. All rights reserved.
         </S.Copyright>
-      </Container>
+        <S.Netlify>
+          This site is powered by
+          <Image
+            src='images/svg/logo-netlify-small-fullcolor-darkmode.svg'
+            width={50}
+            height={50}
+            alt='Netlify Logo'
+          />
+        </S.Netlify>
+      </S.BottomText>
     </S.Footer>
   );
 }