+++ next.config.js
@@ -22,10 +22,10 @@ module.exports = withPWA({
         protocol: 'https',
         hostname: 'res.cloudinary.com',
       },
-      { // Add this new pattern for media.dev.to
+      {
         protocol: 'https',
-        hostname: 'media.dev.to'
-      }
+        hostname: 'media.dev.to',
+      },
     ],
   },
 });