+++ pages/api/validateReCaptcha.js
@@ -1,14 +1,18 @@
+import sendEmail from './sendEmail.js';
+
 export default async function handler(req, res) {
-  const SECRET_KEY = process.env.RECAPTCHA_SECRET_KEY;
-
-  const { body, method } = req;
-
-  // Extract the email, name, and captcha code from the request body
-  const { email, name, captcha } = body;
+  const { method } = req;
 
   if (method === 'POST') {
+    const SECRET_KEY = process.env.RECAPTCHA_SECRET_KEY;
+    // Extract the email, name, and captcha code from the request body
+    // in case it is a call from the ContactUsForm, it will contain
+    // data used to send email, as well
+    const { name, email, subject, message, subscribe, gReCaptchaToken } =
+      req.body;
+
     // If email or captcha are missing return an error
-    if (!email || !name || !captcha) {
+    if (!email || !name || !gReCaptchaToken) {
       return res.status(422).json({
         message: 'Unprocessable request, please provide the required fields',
       });
@@ -17,7 +21,7 @@ export default async function handler(req, res) {
     try {
       // Ping the google recaptcha verify API to verify the captcha code you received
       const response = await fetch(
-        `https://www.google.com/recaptcha/api/siteverify?secret=${SECRET_KEY}&response=${captcha}`,
+        `https://www.google.com/recaptcha/api/siteverify?secret=${SECRET_KEY}&response=${gReCaptchaToken}`,
         {
           headers: {
             'Content-Type': 'application/x-www-form-urlencoded; charset=utf-8',
@@ -26,8 +30,22 @@ export default async function handler(req, res) {
         }
       );
       const captchaValidation = await response.json();
-
       if (captchaValidation.success) {
+        // in case this call is coming from ContactUsForm, it will send an email too
+        if (subject || message) {
+          // it sends the email
+          const sendEmailOK = await sendEmail(
+            email,
+            name,
+            subject,
+            message,
+            subscribe
+          );
+
+          if (sendEmailOK.status !== 'okay') {
+            return res.status(400).json({ message: sendEmailOK.message });
+          }
+        }
         // Return 200 if everything is successful
         return res.status(200).send('OK');
       }